refactor viscript and fix paths,move glRender.go input.go and event processing in hypervisor class
