+++ hypervisor/hypervisor.go
@@ -1,25 +1,229 @@
-package main
+package hypervisor
 
 import (
+	"fmt"
+	"image"
+	"image/draw"
+	_ "image/png"
+	"log"
+	"os"
+	/*
+		"go/build"
+		"runtime"
+	*/
 	"bytes"
 	"encoding/binary"
-	"fmt"
+	"github.com/corpusc/viscript/app"
 	"github.com/corpusc/viscript/gfx"
 	"github.com/corpusc/viscript/msg"
 	"github.com/corpusc/viscript/script"
 	"github.com/corpusc/viscript/ui"
+	"github.com/go-gl/gl/v2.1/gl"
 	"github.com/go-gl/glfw/v3.2/glfw"
 	"math"
+
+
+
+
+	//"github.com/corpusc/viscript/msg" //message structs are here
+
 )
 
 const PREFIX_SIZE = 5 // guaranteed minimum size of every message (4 for length & 1 for type)
-var events = make(chan []byte, 256)
+const (
+	MessageMousePos    = iota // 0
+	MessageMouseScroll        // 1
+	MessageMouseButton        // 2
+	MessageCharacter
+	MessageKey
+)
+
+
+
+var Events = make(chan []byte, 256)
 var prevMousePixelX float64
 var prevMousePixelY float64
 var mousePixelDeltaX float64
 var mousePixelDeltaY float64
 
-func initInputEvents(w *glfw.Window) {
+var (
+	Texture   uint32
+	rotationX float32
+	rotationY float32
+)
+
+func InitRenderer() {
+	fmt.Println("initRenderer()")
+
+	gl.Enable(gl.DEPTH_TEST)
+	gl.Enable(gl.LIGHTING)
+	//gl.Enable(gl.ALPHA_TEST)
+
+	gl.ClearColor(0.5, 0.5, 0.5, 0.0)
+	gl.ClearDepth(1)
+	gl.DepthFunc(gl.LEQUAL)
+
+	ambient := []float32{0.5, 0.5, 0.5, 1}
+	diffuse := []float32{1, 1, 1, 1}
+	lightPosition := []float32{-5, 5, 10, 0}
+	gl.Lightfv(gl.LIGHT0, gl.AMBIENT, &ambient[0])
+	gl.Lightfv(gl.LIGHT0, gl.DIFFUSE, &diffuse[0])
+	gl.Lightfv(gl.LIGHT0, gl.POSITION, &lightPosition[0])
+	gl.Enable(gl.LIGHT0)
+
+	gl.MatrixMode(gl.PROJECTION)
+	gl.LoadIdentity()
+	setFrustum(gfx.InitFrustum)
+	//gl.Frustum(-1, 1, -1, 1, 1.0, 10.0)
+	//gl.Frustum(left, right, bottom, top, zNear, zFar)
+	gl.MatrixMode(gl.MODELVIEW)
+	gl.LoadIdentity()
+
+	// future FIXME: finished app would not have a demo program loaded on startup?
+	script.Process(false)
+}
+
+func setFrustum(r *app.Rectangle) {
+	gl.Frustum(
+		float64(r.Left),
+		float64(r.Right),
+		float64(r.Bottom),
+		float64(r.Top), 1.0, 10.0)
+}
+
+func DrawScene() {
+	//rotationX += 0.5
+	//rotationY += 0.5
+	gl.Viewport(0, 0, gfx.CurrAppWidth, gfx.CurrAppHeight) // OPTIMIZEME?  could set flag upon frame buffer size change event
+	if *gfx.PrevFrustum != *gfx.CurrFrustum {
+		*gfx.PrevFrustum = *gfx.CurrFrustum
+		gl.MatrixMode(gl.PROJECTION)
+		gl.LoadIdentity()
+		setFrustum(gfx.CurrFrustum)
+		fmt.Println("CHANGE OF FRUSTUM")
+		fmt.Printf(".Panels[0].Rect.Right: %.2f\n", gfx.Rend.Panels[0].Rect.Right)
+		fmt.Printf(".Panels[0].Rect.Top: %.2f\n", gfx.Rend.Panels[0].Rect.Top)
+	}
+	gl.MatrixMode(gl.MODELVIEW) //.PROJECTION)                   //.MODELVIEW)
+	gl.LoadIdentity()
+	gl.Clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT)
+	gl.Translatef(0, 0, -gfx.Rend.DistanceFromOrigin)
+	//gl.Rotatef(rotationX, 1, 0, 0)
+	//gl.Rotatef(rotationY, 0, 1, 0)
+
+	gl.BindTexture(gl.TEXTURE_2D, Texture)
+
+	gl.Begin(gl.QUADS)
+	gfx.Rend.DrawAll()
+	gl.End()
+}
+
+func NewTexture(file string) uint32 {
+	imgFile, err := os.Open(file)
+	if err != nil {
+		log.Fatalf("texture %q not found on disk: %v\n", file, err)
+	}
+
+	img, _, err := image.Decode(imgFile)
+	if err != nil {
+		panic(err)
+	}
+
+	rgba := image.NewRGBA(img.Bounds())
+	if rgba.Stride != rgba.Rect.Size().X*4 {
+		panic("unsupported stride")
+	}
+
+	draw.Draw(rgba, rgba.Bounds(), img, image.Point{0, 0}, draw.Src)
+
+	var texture uint32
+	gl.Enable(gl.TEXTURE_2D)
+	gl.GenTextures(1, &texture)
+	gl.BindTexture(gl.TEXTURE_2D, texture)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE)
+	gl.TexImage2D(
+		gl.TEXTURE_2D,
+		0,
+		gl.RGBA,
+		int32(rgba.Rect.Size().X),
+		int32(rgba.Rect.Size().Y),
+		0,
+		gl.RGBA,
+		gl.UNSIGNED_BYTE,
+		gl.Ptr(rgba.Pix))
+
+	return texture
+}
+
+func destroyScene() {
+}
+
+func makeCube() {
+	gl.Normal3f(0, 0, 1)
+	gl.TexCoord2f(0, 0)
+	gl.Vertex3f(-1, -1, 1)
+	gl.TexCoord2f(1, 0)
+	gl.Vertex3f(1, -1, 1)
+	gl.TexCoord2f(1, 1)
+	gl.Vertex3f(1, 1, 1)
+	gl.TexCoord2f(0, 1)
+	gl.Vertex3f(-1, 1, 1)
+
+	gl.Normal3f(0, 0, -1)
+	gl.TexCoord2f(1, 0)
+	gl.Vertex3f(-1, -1, -1)
+	gl.TexCoord2f(1, 1)
+	gl.Vertex3f(-1, 1, -1)
+	gl.TexCoord2f(0, 1)
+	gl.Vertex3f(1, 1, -1)
+	gl.TexCoord2f(0, 0)
+	gl.Vertex3f(1, -1, -1)
+
+	gl.Normal3f(0, 1, 0)
+	gl.TexCoord2f(0, 1)
+	gl.Vertex3f(-1, 1, -1)
+	gl.TexCoord2f(0, 0)
+	gl.Vertex3f(-1, 1, 1)
+	gl.TexCoord2f(1, 0)
+	gl.Vertex3f(1, 1, 1)
+	gl.TexCoord2f(1, 1)
+	gl.Vertex3f(1, 1, -1)
+
+	gl.Normal3f(0, -1, 0)
+	gl.TexCoord2f(1, 1)
+	gl.Vertex3f(-1, -1, -1)
+	gl.TexCoord2f(0, 1)
+	gl.Vertex3f(1, -1, -1)
+	gl.TexCoord2f(0, 0)
+	gl.Vertex3f(1, -1, 1)
+	gl.TexCoord2f(1, 0)
+	gl.Vertex3f(-1, -1, 1)
+
+	gl.Normal3f(1, 0, 0)
+	gl.TexCoord2f(1, 0)
+	gl.Vertex3f(1, -1, -1)
+	gl.TexCoord2f(1, 1)
+	gl.Vertex3f(1, 1, -1)
+	gl.TexCoord2f(0, 1)
+	gl.Vertex3f(1, 1, 1)
+	gl.TexCoord2f(0, 0)
+	gl.Vertex3f(1, -1, 1)
+
+	gl.Normal3f(-1, 0, 0)
+	gl.TexCoord2f(0, 0)
+	gl.Vertex3f(-1, -1, -1)
+	gl.TexCoord2f(1, 0)
+	gl.Vertex3f(-1, -1, 1)
+	gl.TexCoord2f(1, 1)
+	gl.Vertex3f(-1, 1, 1)
+	gl.TexCoord2f(0, 1)
+	gl.Vertex3f(-1, 1, -1)
+}
+
+func InitInputEvents(w *glfw.Window) {
 	w.SetCharCallback(onChar)
 	w.SetKeyCallback(onKey)
 	w.SetMouseButtonCallback(onMouseButton)
@@ -162,12 +366,12 @@ func onKey(
 		case glfw.KeyLeftControl:
 			fallthrough
 		case glfw.KeyRightControl:
-			//fmt.Println("Control RELEASED")
+		//fmt.Println("Control RELEASED")
 
 		case glfw.KeyLeftAlt:
 			fallthrough
 		case glfw.KeyRightAlt:
-			//fmt.Println("Alt RELEASED")
+		//fmt.Println("Alt RELEASED")
 
 		case glfw.KeyLeftSuper:
 			fallthrough
@@ -286,7 +490,7 @@ func dispatchWithPrefix(content []byte, msgType uint8) {
 		getBytesOfUInt32(uint32(len(content))+PREFIX_SIZE),
 		getByteOfUInt8(msgType)...)
 
-	events <- append(prefix, content...)
+	Events <- append(prefix, content...)
 }
 
 func getWordSkipPos(xIn int, change int) int {
@@ -375,4 +579,195 @@ func getSlice(wBuf *bytes.Buffer, err error) (data []byte) {
 	}
 
 	return
+}
+
+
+
+var curRecByte = 0 // current receive message index
+
+func monitorEvents(ch chan []byte) {
+	select {
+	case v := <-ch:
+		processMessage(v)
+	default:
+	//fmt.Println("monitorEvents() default")
+	}
+}
+
+/* message processing example */
+
+/*
+func ProcessIncomingMessages() {
+	//have a channel for incoming messages
+	for msg := range self.IncomingChannel {
+		switch msg.GetMessageType(msg) {
+		//InRouteMessage is the only message coming in to node from transports
+		case msg.MsgMessageMousePos:
+			var m1 msg.MessageMousePos
+			msg.Deserialize(msg, m1)
+			//self.HandleInRouteMessage(m1)
+			fmt.Printf("MessageMousePos: X= %f, Y= %f \n", m1.X, m2.X)
+		case msg.MsgMessageMouseScroll:
+			var m2 msg.MessageMouseScroll
+			msg.Deserialize(msg, m1)
+
+		case msg.MsgMessageMouseButton:
+			var m3 msg.MessageMouseButton
+			mesg.Deserialize(msg, m1)
+
+		case msg.MsgMessageCharacter:
+			var m4 msg.MessageCharacter
+			msg.Deserialize(msg, m1)
+
+		case msg.MsgMessageKey:
+			var m5 msg.MessageKey
+			msg.Deserialize(msg, m1)
+
+		default:
+			fmt.Println("UNKNOWN MESSAGE TYPE!")
+
+		}
+	}
+}
+*/
+
+func processMessage(message []byte) {
+	switch getMessageType(".", message) {
+
+	case MessageMousePos:
+		s("MessageMousePos", message)
+		showFloat64("X", message)
+		showFloat64("Y", message)
+
+	case MessageMouseScroll:
+		s("MessageMouseScroll", message)
+		showFloat64("X Offset", message)
+		showFloat64("Y Offset", message)
+
+	case MessageMouseButton:
+		s("MessageMouseButton", message)
+		gfx.Curs.ConvertMouseClickToTextCursorPosition(
+			getAndShowUInt8("Button", message),
+			getAndShowUInt8("Action", message))
+		getAndShowUInt8("Mod", message)
+
+	case MessageCharacter:
+		s("MessageCharacter", message)
+		insertRuneIntoDocument("Rune", message)
+		script.Process(false)
+
+	case MessageKey:
+		s("MessageKey", message)
+		getAndShowUInt8("Key", message)
+		showSInt32("Scan", message)
+		getAndShowUInt8("Action", message)
+		getAndShowUInt8("Mod", message)
+
+	default:
+		fmt.Println("UNKNOWN MESSAGE TYPE!")
+	}
+
+	fmt.Println()
+	curRecByte = 0
+}
+
+func s(s string, message []byte) {
+	fmt.Print(s)
+	showUInt32("Len", message)
+	curRecByte++ // skipping message type's space
+}
+
+func getMessageType(s string, message []byte) (value uint8) {
+	rBuf := bytes.NewReader(message[4:5])
+	err := binary.Read(rBuf, binary.LittleEndian, &value)
+
+	if err != nil {
+		fmt.Println("binary.Read failed: ", err)
+	} else {
+		//fmt.Printf("from byte buffer, %s: %d\n", s, value)
+	}
+
+	return
+}
+
+func insertRuneIntoDocument(s string, message []byte) {
+	var value rune
+	var size = 4
+
+	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
+	err := binary.Read(rBuf, binary.LittleEndian, &value)
+	curRecByte += size
+
+	if err != nil {
+		fmt.Println("binary.Read failed: ", err)
+	} else {
+		fmt.Printf("   [%s: %s]", s, string(value))
+
+		f := gfx.Rend.Focused
+		b := f.TextBodies[0]
+		b[f.CursY] = b[f.CursY][:f.CursX] + string(value) + b[f.CursY][f.CursX:len(b[f.CursY])]
+		f.CursX++
+	}
+}
+
+func getAndShowUInt8(s string, message []byte) (value uint8) {
+	var size = 1
+
+	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
+	err := binary.Read(rBuf, binary.LittleEndian, &value)
+	curRecByte += size
+
+	if err != nil {
+		fmt.Println("binary.Read failed: ", err)
+	} else {
+		fmt.Printf("   [%s: %d]", s, value)
+	}
+
+	return
+}
+
+// the rest of these funcs are almost identical, just top 2 vars customized (and string format)
+func showSInt32(s string, message []byte) {
+	var value int32
+	var size = 4
+
+	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
+	err := binary.Read(rBuf, binary.LittleEndian, &value)
+	curRecByte += size
+
+	if err != nil {
+		fmt.Println("binary.Read failed: ", err)
+	} else {
+		fmt.Printf("   [%s: %d]", s, value)
+	}
+}
+
+func showUInt32(s string, message []byte) {
+	var value uint32
+	var size = 4
+
+	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
+	err := binary.Read(rBuf, binary.LittleEndian, &value)
+	curRecByte += size
+
+	if err != nil {
+		fmt.Println("binary.Read failed: ", err)
+	} else {
+		fmt.Printf("   [%s: %d]", s, value)
+	}
+}
+
+func showFloat64(s string, message []byte) {
+	var value float64
+	var size = 8
+
+	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
+	err := binary.Read(rBuf, binary.LittleEndian, &value)
+	curRecByte += size
+
+	if err != nil {
+		fmt.Println("binary.Read failed: ", err)
+	} else {
+		fmt.Printf("   [%s: %.1f]", s, value)
+	}
 }
