--- messages.go
@@ -1,207 +1,0 @@
-package main
-
-import (
-	"bytes"
-	"encoding/binary"
-	"fmt"
-	//"github.com/corpusc/viscript/msg" //message structs are here
-	"github.com/corpusc/viscript/gfx"
-	"github.com/corpusc/viscript/script"
-)
-
-const (
-	MessageMousePos    = iota // 0
-	MessageMouseScroll        // 1
-	MessageMouseButton        // 2
-	MessageCharacter
-	MessageKey
-)
-
-var curRecByte = 0 // current receive message index
-
-func monitorEvents(ch chan []byte) {
-	select {
-	case v := <-ch:
-		processMessage(v)
-	default:
-		//fmt.Println("monitorEvents() default")
-	}
-}
-
-/* message processing example */
-
-/*
-func ProcessIncomingMessages() {
-	//have a channel for incoming messages
-	for msg := range self.IncomingChannel {
-		switch msg.GetMessageType(msg) {
-		//InRouteMessage is the only message coming in to node from transports
-		case msg.MsgMessageMousePos:
-			var m1 msg.MessageMousePos
-			msg.Deserialize(msg, m1)
-			//self.HandleInRouteMessage(m1)
-			fmt.Printf("MessageMousePos: X= %f, Y= %f \n", m1.X, m2.X)
-		case msg.MsgMessageMouseScroll:
-			var m2 msg.MessageMouseScroll
-			msg.Deserialize(msg, m1)
-
-		case msg.MsgMessageMouseButton:
-			var m3 msg.MessageMouseButton
-			mesg.Deserialize(msg, m1)
-
-		case msg.MsgMessageCharacter:
-			var m4 msg.MessageCharacter
-			msg.Deserialize(msg, m1)
-
-		case msg.MsgMessageKey:
-			var m5 msg.MessageKey
-			msg.Deserialize(msg, m1)
-
-		default:
-			fmt.Println("UNKNOWN MESSAGE TYPE!")
-
-		}
-	}
-}
-*/
-
-func processMessage(message []byte) {
-	switch getMessageType(".", message) {
-
-	case MessageMousePos:
-		s("MessageMousePos", message)
-		showFloat64("X", message)
-		showFloat64("Y", message)
-
-	case MessageMouseScroll:
-		s("MessageMouseScroll", message)
-		showFloat64("X Offset", message)
-		showFloat64("Y Offset", message)
-
-	case MessageMouseButton:
-		s("MessageMouseButton", message)
-		gfx.Curs.ConvertMouseClickToTextCursorPosition(
-			getAndShowUInt8("Button", message),
-			getAndShowUInt8("Action", message))
-		getAndShowUInt8("Mod", message)
-
-	case MessageCharacter:
-		s("MessageCharacter", message)
-		insertRuneIntoDocument("Rune", message)
-		script.Process(false)
-
-	case MessageKey:
-		s("MessageKey", message)
-		getAndShowUInt8("Key", message)
-		showSInt32("Scan", message)
-		getAndShowUInt8("Action", message)
-		getAndShowUInt8("Mod", message)
-
-	default:
-		fmt.Println("UNKNOWN MESSAGE TYPE!")
-	}
-
-	fmt.Println()
-	curRecByte = 0
-}
-
-func s(s string, message []byte) {
-	fmt.Print(s)
-	showUInt32("Len", message)
-	curRecByte++ // skipping message type's space
-}
-
-func getMessageType(s string, message []byte) (value uint8) {
-	rBuf := bytes.NewReader(message[4:5])
-	err := binary.Read(rBuf, binary.LittleEndian, &value)
-
-	if err != nil {
-		fmt.Println("binary.Read failed: ", err)
-	} else {
-		//fmt.Printf("from byte buffer, %s: %d\n", s, value)
-	}
-
-	return
-}
-
-func insertRuneIntoDocument(s string, message []byte) {
-	var value rune
-	var size = 4
-
-	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
-	err := binary.Read(rBuf, binary.LittleEndian, &value)
-	curRecByte += size
-
-	if err != nil {
-		fmt.Println("binary.Read failed: ", err)
-	} else {
-		fmt.Printf("   [%s: %s]", s, string(value))
-
-		f := gfx.Rend.Focused
-		b := f.TextBodies[0]
-		b[f.CursY] = b[f.CursY][:f.CursX] + string(value) + b[f.CursY][f.CursX:len(b[f.CursY])]
-		f.CursX++
-	}
-}
-
-func getAndShowUInt8(s string, message []byte) (value uint8) {
-	var size = 1
-
-	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
-	err := binary.Read(rBuf, binary.LittleEndian, &value)
-	curRecByte += size
-
-	if err != nil {
-		fmt.Println("binary.Read failed: ", err)
-	} else {
-		fmt.Printf("   [%s: %d]", s, value)
-	}
-
-	return
-}
-
-// the rest of these funcs are almost identical, just top 2 vars customized (and string format)
-func showSInt32(s string, message []byte) {
-	var value int32
-	var size = 4
-
-	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
-	err := binary.Read(rBuf, binary.LittleEndian, &value)
-	curRecByte += size
-
-	if err != nil {
-		fmt.Println("binary.Read failed: ", err)
-	} else {
-		fmt.Printf("   [%s: %d]", s, value)
-	}
-}
-
-func showUInt32(s string, message []byte) {
-	var value uint32
-	var size = 4
-
-	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
-	err := binary.Read(rBuf, binary.LittleEndian, &value)
-	curRecByte += size
-
-	if err != nil {
-		fmt.Println("binary.Read failed: ", err)
-	} else {
-		fmt.Printf("   [%s: %d]", s, value)
-	}
-}
-
-func showFloat64(s string, message []byte) {
-	var value float64
-	var size = 8
-
-	rBuf := bytes.NewReader(message[curRecByte : curRecByte+size])
-	err := binary.Read(rBuf, binary.LittleEndian, &value)
-	curRecByte += size
-
-	if err != nil {
-		fmt.Println("binary.Read failed: ", err)
-	} else {
-		fmt.Printf("   [%s: %.1f]", s, value)
-	}
-}
