+++ viscript.go
@@ -41,6 +41,16 @@ import (
 	"github.com/corpusc/viscript/msg"
 	"github.com/go-gl/gl/v2.1/gl"
 	"github.com/go-gl/glfw/v3.2/glfw"
+	"github.com/corpusc/viscript/hypervisor"
+
+	//	"log"
+	//	"os"
+	//	/*
+	//		"go/build"
+	//		"runtime"
+	//	*/
+
+
 )
 
 func init() {
@@ -72,17 +82,20 @@ func main() {
 	if err := gl.Init(); err != nil {
 		panic(err)
 	}
+	hypervisor.Texture = hypervisor.NewTexture("Bisasam_24x24_Shadowed.png")
 
-	texture = newTexture("Bisasam_24x24_Shadowed.png")
-	defer gl.DeleteTextures(1, &texture)
+	defer gl.DeleteTextures(1, &hypervisor.Texture)
 
-	initRenderer()
-	initInputEvents(window)
+	hypervisor.InitRenderer()
+
+	hypervisor.InitInputEvents(window)
+
 
 	for !window.ShouldClose() {
-		msg.MonitorEvents(events)
+		msg.MonitorEvents(hypervisor.Events)
 		glfw.PollEvents()
-		drawScene()
+		hypervisor.DrawScene()
+		//drawScene()
 		window.SwapBuffers()
 	}
 }
